[K8s] Move integration tests to own folder
