+++ kubernetes/tests/integration/integration/test_exporter_with_tls.py
@@ -12,7 +12,7 @@ import tenacity
 import yaml
 from pytest_operator.plugin import OpsTest
 
-from .helpers import (
+from ..helpers import (
     APPLICATION_DEFAULT_APP_NAME,
     MYSQL_DEFAULT_APP_NAME,
     MYSQL_ROUTER_DEFAULT_APP_NAME,
