+++ kubernetes/tests/integration/integration/test_tls.py
@@ -10,7 +10,7 @@ import tenacity
 import yaml
 from pytest_operator.plugin import OpsTest
 
-from .helpers import (
+from ..helpers import (
     APPLICATION_DEFAULT_APP_NAME,
     MYSQL_DEFAULT_APP_NAME,
     MYSQL_ROUTER_DEFAULT_APP_NAME,
